+++ hosts/apps1.nix
@@ -171,6 +171,7 @@
       autoStart = true;
       hostname = "forgejo";
       image = "code.forgejo.org/forgejo/forgejo:14";
+      environment = config.secrets.forgejo.keys;
       networks = [
         "backend"
       ];